rename private headers accordingly and include them as so in CMakeLists
