+++ src/core/maprenderer/qgsmaprenderercache.cpp
@@ -16,7 +16,7 @@
 #include "qgsmaprenderercache.h"
 
 #include "qgsmaplayer.h"
-#include "qgsmaplayerlistutils.h"
+#include "qgsmaplayerlistutils_p.h"
 #include "qgsapplication.h"
 
 #include <QImage>
