+++ src/core/maprenderer/qgsmaprenderercustompainterjob.cpp
@@ -19,7 +19,7 @@
 #include "qgslabelingengine.h"
 #include "qgslogger.h"
 #include "qgsmaplayerrenderer.h"
-#include "qgsmaplayerlistutils.h"
+#include "qgsmaplayerlistutils_p.h"
 #include "qgsvectorlayerlabeling.h"
 
 #include <QtConcurrentRun>
