+++ src/core/processing/qgsprocessingcontext.cpp
@@ -18,7 +18,7 @@
 #include "qgsprocessingcontext.h"
 #include "qgsprocessingutils.h"
 #include "qgsproviderregistry.h"
-#include "qgsmaplayerlistutils.h"
+#include "qgsmaplayerlistutils_p.h"
 #include "qgssettings.h"
 
 QgsProcessingContext::QgsProcessingContext()
