+++ src/core/proj/qgscoordinatereferencesystem_legacy_p.h
