+++ src/core/providers/ogr/qgscplerrorhandler_p.h
