+++ src/core/qgsmaplayerlistutils_p.h
