+++ tests/src/core/testqgsmapsettings.cpp
@@ -25,7 +25,7 @@
 #include "qgspointxy.h"
 #include "qgslogger.h"
 #include "qgsapplication.h"
-#include "qgsmaplayerlistutils.h"
+#include "qgsmaplayerlistutils_p.h"
 #include "qgsvectorlayer.h"
 #include "qgscoordinatereferencesystem.h"
 #include "qgsexpressioncontextutils.h"
